requirements.txt
python-dotenv
PyMuPDF
faiss-cpu
langchain
langchain-community
langchain-openai